+++ src/xPLServer.h
@@ -8,6 +8,8 @@
 #ifndef __xPLServer_h
 #define __xPLServer_h
 
+#include <inttypes.h>
+
 #include "queue.h"
 #include "xPL.h"
 
@@ -15,7 +17,7 @@ extern char *xpl_vendorID;
 extern char *xpl_deviceID;
 extern char *xpl_instanceID;
 
-typedef int (*xpl_f)(xPL_ServicePtr,xPL_MessagePtr theMessage, xPL_ObjectPtr userValue);
+typedef int16_t (*xpl_f)(xPL_ServicePtr,xPL_MessagePtr theMessage, xPL_ObjectPtr userValue);
 
 pthread_t *xPLServer(queue_t *interfaces);
 xPL_ServicePtr get_xPL_ServicePtr();
